requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
flask>=3.0.0
anthropic>=0.40.0
pytest>=7.4.0
responses>=0.24.0
